Implement logic to recover a frame exit_code
When rqd restarts all frames currently running become parented by
the init/systemd, therefore the recovery process is no longer able
to bind to them and follow their execution directly.

To accomplish a similar behaviour, on recovery mode, running_frame
pools it's pid waiting for it to finish and recovers the exit status
and signal from a file. To accomplish that, all frame commands are
now appended with a logic to save their exit status to a file.